+++ services/frontend/src/app/layout/digital-services-footprint/digital-services-footprint-header/digital-services-footprint-header.component.ts
@@ -70,6 +70,12 @@ export class DigitalServicesFootprintHeaderComponent implements OnInit {
     digitalServiceVersionUid =
         this.route.snapshot.paramMap.get("digitalServiceVersionId") ?? "";
     isPromoteVersionDialogVisible = false;
+    duplicateDsNames: string[] = [];
+    duplicateVersionNames: string[] = [];
+    disableDs = false;
+    disableVersion = false;
+    savedDigitalServiceAndVersion: any = undefined;
+    firstDsVersionCall = true;
     private readonly destroyRef = inject(DestroyRef);
 
     constructor(
@@ -83,6 +89,7 @@ export class DigitalServicesFootprintHeaderComponent implements OnInit {
     ) {}
 
     ngOnInit() {
+        console.log("in1");
         this.route.paramMap
             .pipe(takeUntilDestroyed(this.destroyRef))
             .subscribe((params) => {
@@ -94,6 +101,10 @@ export class DigitalServicesFootprintHeaderComponent implements OnInit {
             .pipe(takeUntilDestroyed(this.destroyRef))
             .subscribe((res) => {
                 this.digitalService = res;
+                this.savedDigitalServiceAndVersion = {
+                    dsName: this.digitalService.name,
+                    version: this.digitalService.description,
+                };
                 this.digitalServiceStore.setDigitalService(this.digitalService);
             });
 
@@ -125,6 +136,7 @@ export class DigitalServicesFootprintHeaderComponent implements OnInit {
                 this.digitalService.description = digitalServiceName;
             }
             this.digitalServiceChange.emit(this.digitalService);
+            this.firstDsVersionCall = true;
         }
     }
 
@@ -285,5 +297,54 @@ export class DigitalServicesFootprintHeaderComponent implements OnInit {
                     );
                 }
             });
+    }
+
+    callDSVersion(value: string, isDs: boolean) {
+        this.digitalServicesData
+            .getDuplicateDigitalServiceAndVersionName(this.digitalServiceVersionUid)
+            .pipe(takeUntilDestroyed(this.destroyRef))
+            .subscribe((n) => {
+                this.duplicateDsNames = n.dsNames;
+                this.duplicateVersionNames = n.versionNames;
+                if (isDs) {
+                    this.validateDataDS(value);
+                } else {
+                    this.validateDataVersion(value);
+                }
+                this.firstDsVersionCall = false;
+            });
+    }
+
+    validateDataDS(value: string) {
+        this.disableDs = this.duplicateDsNames
+            .map((ds) => ds.trim())
+            .filter(
+                (ds) => ds.trim() !== this.savedDigitalServiceAndVersion.dsName.trim(),
+            )
+            .includes(value.trim());
+    }
+
+    validateDataVersion(value: string) {
+        console.log(this.savedDigitalServiceAndVersion);
+        this.disableVersion = this.duplicateVersionNames
+            .map((v) => v.trim())
+            .filter((v) => v.trim() !== this.savedDigitalServiceAndVersion.version.trim())
+            .includes(value.trim());
+    }
+
+    validateDs(value: string) {
+        if (this.firstDsVersionCall) {
+            this.callDSVersion(value, true);
+        } else {
+            this.validateDataDS(value);
+        }
+    }
+
+    validateVersion(value: string) {
+        if (this.firstDsVersionCall) {
+            this.callDSVersion(value, false);
+        } else {
+            this.validateDataVersion(value);
+        }
     }
 }